Wired up react
Removed static content. Established react.js as the entry point for our
React app and separated phoenix_html out to it's own director.
