+++ apps/connect_four_backend/lib/connect_four_backend_web/static/js/application.js
@@ -1,15 +1,2 @@
-'use strict';
-
-var elements = document.querySelectorAll('[data-submit^=parent]');
-var len = elements.length;
-
-for (var i = 0; i < len; ++i) {
-  elements[i].addEventListener('click', function (event) {
-    var message = this.getAttribute("data-confirm");
-    if (message === null || confirm(message)) {
-      this.parentNode.submit();
-    };
-    event.preventDefault();
-    return false;
-  }, false);
-};
+import './phoenix_html';
+import './react';
